Refactor structopt cli into multiple files
